+++ beaker/data_model/cluster.py
@@ -40,6 +40,7 @@ class Cluster(BaseModel):
     validated: Optional[datetime] = None
     user_restrictions: Optional[Tuple[str, ...]] = None
     allow_preemptible_restriction_exceptions: Optional[bool] = None
+    compute_source: Optional[str] = None
 
     @field_validator("validated")
     def _validate_datetime(cls, v: Optional[datetime]) -> Optional[datetime]: